переносит константы в отдельный модуль, функцию рендера в утилс
