+++ src/constants.js
@@ -1,0 +1,28 @@
+const ExtraTitles = {
+  TOP_RATED: `Top rated`,
+  MOST_COMMENTED: `Most commented`
+};
+
+const TypeOfSorting = {
+  rating: `rating`,
+  comments: `comments`
+};
+
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+const Count = {
+  CARD: 29,
+  SHOWING_CARDS_ON_START: 5,
+  SHOWING_CARDS_BY_BUTTON: 5,
+  EXTRA: 2,
+  EXTRA_FILMS: 2
+};
+
+const KeyCode = {
+  ESC: 27
+};
+
+export {ExtraTitles, TypeOfSorting, RenderPosition, Count, KeyCode};
